requirements.txt
Flask>=2.3.2
requests==2.31.0
Werkzeug>=2.2,<3.0
Flask_Cors==4.0.1
python-dotenv==1.0.1